+++ src/main/java/com/aquent/crudapp/controllers/ClientController.java
@@ -1,7 +1,8 @@
-package com.aquent.crudapp.client;
+package com.aquent.crudapp.controllers;
 
+import com.aquent.crudapp.model.client.Client;
 import com.aquent.crudapp.interfaces.EntityService;
-import com.aquent.crudapp.person.Person;
+import com.aquent.crudapp.model.person.Person;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
