+++ src/main/java/com/aquent/crudapp/controllers/HomeController.java
@@ -1,4 +1,4 @@
-package com.aquent.crudapp;
+package com.aquent.crudapp.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
