+++ src/main/java/com/aquent/crudapp/interfaces/Entity.java
@@ -4,20 +4,6 @@ public interface Entity {
 
     Integer getEntityId();
 
-    String getStreetAddress();
-
-    void setStreetAddress(String streetAddress);
-
-    String getCity();
-
-    void setCity(String city);
-
-    String getState();
-
-    void setState(String state);
-
-    String getZipCode();
-
-    void setZipCode(String zipCode);
+    void setEntityId(Integer entityId);
 
 }
