+++ src/main/java/com/aquent/crudapp/model/AbstractEntityWithAddress.java
@@ -1,11 +1,11 @@
-package com.aquent.crudapp.abstracts;
+package com.aquent.crudapp.model;
 
 import com.aquent.crudapp.interfaces.Entity;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
-public class EntityAbstract implements Entity {
+public class AbstractEntityWithAddress implements Entity {
 
     private Integer entityId;
