+++ src/main/java/com/aquent/crudapp/services/ClientService.java
@@ -1,8 +1,9 @@
-package com.aquent.crudapp.client;
+package com.aquent.crudapp.services;
 
+import com.aquent.crudapp.model.client.Client;
 import com.aquent.crudapp.interfaces.EntityDao;
 import com.aquent.crudapp.interfaces.EntityService;
-import com.aquent.crudapp.person.Person;
+import com.aquent.crudapp.model.person.Person;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
