+++ src/main/java/com/aquent/crudapp/services/PersonService.java
@@ -1,4 +1,4 @@
-package com.aquent.crudapp.person;
+package com.aquent.crudapp.services;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -7,9 +7,10 @@ import java.util.Set;
 import javax.validation.ConstraintViolation;
 import javax.validation.Validator;
 
-import com.aquent.crudapp.client.Client;
+import com.aquent.crudapp.model.client.Client;
 import com.aquent.crudapp.interfaces.EntityDao;
 import com.aquent.crudapp.interfaces.EntityService;
+import com.aquent.crudapp.model.person.Person;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
